pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-reliability-framework"
version = "2.0.0"
description = "Production-grade multi-agent AI system for infrastructure reliability monitoring and self-healing"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Juan Petter", email = "petter2025us@outlook.com"}
]
keywords = [
    "ai-agents",
    "reliability-engineering",
    "mlops",
    "observability",
    "self-healing",
    "anomaly-detection",
    "predictive-analytics"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

dependencies = [
    "gradio>=5.0.0,<6.0.0",
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "requests>=2.32.5",
    "circuitbreaker>=1.4.0",
    "atomicwrites>=1.4.1",
    "python-dotenv>=1.0.0",
    "click>=8.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0"
]

[project.urls]
Homepage = "https://github.com/petterjuan/agentic-reliability-framework"
Documentation = "https://github.com/petterjuan/agentic-reliability-framework/blob/main/README.md"
Repository = "https://github.com/petterjuan/agentic-reliability-framework"
Issues = "https://github.com/petterjuan/agentic-reliability-framework/issues"

[project.scripts]
arf = "agentic_reliability_framework.cli:main"

[tool.setuptools]
packages.find = {where = ["."]}
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.json", "*.index", "*.txt", "*.md"]
"agentic_reliability_framework" = ["data/*", "py.typed"]
